Выполнено разделение кода на модули. п.5.12 Модуляция
